Add new packages to plasma-mobile-gear
At least for now, we're excluding khealthcertificate and
vakzination -- people should have the choice of whether or
not to take an experimental treatment without being excluded
from society, and people have a right to privacy (not letting
everyone know what decision they made).

Having to present a certificate on every other street corner
is not compatible with a democratic society, even if the
application used to present it is Open Source.